+++ src/dune_engine/sandbox.ml
@@ -209,9 +209,8 @@ let apply_changes_to_source_tree t ~old_snapshot =
   in
   let copy_file p =
     let in_source_tree = in_source_tree p in
-    Path.unlink_no_err in_source_tree;
     Option.iter (Path.parent in_source_tree) ~f:Path.mkdir_p;
-    Io.copy_file ~src:p ~dst:in_source_tree ()
+    Io.copy_file ~atomically:true ~src:p ~dst:in_source_tree ()
   in
   let delete_file p =
     let in_source_tree = in_source_tree p in